requirements.txt
anthropic==0.25.1
flask==2.0.1
openai==1.17.1
python-dotenv==1.0.1
werkzeug==3.0.2
